LPS-120039 Add confirmation message for field deletion.

+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/components/LayoutProvider/LayoutProvider.es.js
@@ -334,8 +334,17 @@ class LayoutProvider extends Component {
 				label: Liferay.Language.get('duplicate'),
 			},
 			{
-				action: ({activePage, fieldName}) =>
-					this.dispatch('fieldDeleted', {activePage, fieldName}),
+				action: ({activePage, fieldName}) => {
+					if (
+						confirm(
+							Liferay.Language.get(
+								'are-you-sure-you-want-to-delete-this-field'
+							)
+						)
+					) {
+						this.dispatch('fieldDeleted', {activePage, fieldName});
+					}
+				},
 				label: Liferay.Language.get('delete'),
 			},
 		];
